docs: Fixing strict controls
