+++ docs/src/pages/reference/strict-controls/active.astro
@@ -2,7 +2,7 @@
 import StarlightPage from '@astrojs/starlight/components/StarlightPage.astro';
 import StrictControlEntries from '@components/StrictControlEntries.astro';
 import { getCollection } from 'astro:content';
-import { sidebar } from '../../../../astro.config.mjs';
+import { sidebar } from '../../../data/sidebar';
 
 const showUnreleased = Astro.url.searchParams.get('unreleased') === 'true';
